refactor: Rename 'Engine' folder to 'Core'.
