+++ src/StardustDefender/Background/Layers/SBackgroundLayer.cs
@@ -3,7 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 
 using StardustDefender.Camera;
 using StardustDefender.Controllers;
-using StardustDefender.Engine;
+using StardustDefender.Core;
 
 namespace StardustDefender.Background.Layers
 {
