+++ src/StardustDefender/Controllers/Difficulty/SDifficultyController.cs
@@ -1,4 +1,4 @@
-﻿using StardustDefender.Engine;
+﻿using StardustDefender.Core;
 
 namespace StardustDefender.Controllers
 {
