+++ src/StardustDefender/Core/SGraphics.cs
@@ -1,7 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
 
-namespace StardustDefender.Engine
+namespace StardustDefender.Core
 {
     internal static class SGraphics
     {
