+++ src/StardustDefender/Core/SSongs.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace StardustDefender.Engine
+namespace StardustDefender.Core
 {
     internal static class SSongs
     {
