+++ src/StardustDefender/Core/STextures.cs
@@ -3,7 +3,7 @@ using Microsoft.Xna.Framework.Graphics;
 
 using System.Collections.Generic;
 
-namespace StardustDefender.Engine
+namespace StardustDefender.Core
 {
     internal static class STextures
     {
