+++ src/StardustDefender/Entities/Aliens/SAlienEntity.cs
@@ -2,7 +2,7 @@
 
 using StardustDefender.Controllers;
 using StardustDefender.Effects.Common;
-using StardustDefender.Engine;
+using StardustDefender.Core;
 using StardustDefender.Managers;
 
 using System.Threading.Tasks;
