+++ src/StardustDefender/Entities/Aliens/SAlien_01.cs
@@ -1,4 +1,4 @@
-﻿using StardustDefender.Engine;
+﻿using StardustDefender.Core;
 using StardustDefender.Enums;
 
 namespace StardustDefender.Entities.Aliens
