+++ src/StardustDefender/Entities/Aliens/SAlien_02.cs
@@ -1,4 +1,4 @@
-﻿using StardustDefender.Engine;
+﻿using StardustDefender.Core;
 
 using StardustDefender.Enums;
 using StardustDefender.Managers;
