+++ src/StardustDefender/Entities/Bosses/SBossEntity.cs
@@ -1,7 +1,7 @@
 ﻿using StardustDefender.Entities.Player;
 using StardustDefender.Controllers;
 using StardustDefender.Effects.Common;
-using StardustDefender.Engine;
+using StardustDefender.Core;
 using StardustDefender.Managers;
 
 using System.Threading.Tasks;
