+++ src/StardustDefender/Entities/Bosses/SBoss_01.cs
@@ -1,6 +1,6 @@
 ﻿using StardustDefender.Controllers;
 using StardustDefender.Effects.Common;
-using StardustDefender.Engine;
+using StardustDefender.Core;
 using StardustDefender.Managers;
 using StardustDefender.Enums;
 using StardustDefender.Animation;
