+++ src/StardustDefender/Extensions/IEnumerableExtensions.cs
@@ -1,4 +1,4 @@
-﻿using StardustDefender.Engine;
+﻿using StardustDefender.Core;
 
 using System.Collections.Generic;
 using System.Linq;
