+++ src/StardustDefender/Items/Common/SShootSpeedUpgradeItem.cs
@@ -1,4 +1,4 @@
-﻿using StardustDefender.Engine;
+﻿using StardustDefender.Core;
 using StardustDefender.Entities.Player;
 
 namespace StardustDefender.Items.Common
