+++ src/StardustDefender/Items/SItem.cs
@@ -5,7 +5,7 @@ using StardustDefender.Animation;
 using StardustDefender.Collections;
 using StardustDefender.Controllers;
 using StardustDefender.Effects.Common;
-using StardustDefender.Engine;
+using StardustDefender.Core;
 using StardustDefender.Enums;
 using StardustDefender.Managers;
